korean_making/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
api==0.0.7
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
azure-cognitiveservices-speech==1.33.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
dj-rest-auth==5.0.1
Django==4.2.6
django-blacklist==0.7.0
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==2.0.0
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==6.0.4
nose==1.3.7
numba==0.58.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
stack-data==0.6.3
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Wave==0.0.2
wcwidth==0.2.8
yarl==1.9.2
